@@ -60,54 +60,54 @@
     "cypress:run:ci:parameterized": "node update-config.js TESTNET_ENABLED=true TESTNET4_ENABLED=true SIGNET_ENABLED=true LIQUID_ENABLED=true ITEMS_PER_PAGE=25 && npm run generate-config && start-server-and-test serve:parameterized 4200 cypress:run:record"
   },
   "dependencies": {
-    "@angular-devkit/build-angular": "^17.3.1",
-    "@angular/animations": "^17.3.1",
-    "@angular/cli": "^17.3.1",
-    "@angular/common": "^17.3.1",
-    "@angular/compiler": "^17.3.1",
-    "@angular/core": "^17.3.1",
-    "@angular/forms": "^17.3.1",
-    "@angular/localize": "^17.3.1",
-    "@angular/platform-browser": "^17.3.1",
-    "@angular/platform-browser-dynamic": "^17.3.1",
-    "@angular/platform-server": "^17.3.1",
-    "@angular/router": "^17.3.1",
-    "@angular/ssr": "^17.3.1",
-    "@fortawesome/angular-fontawesome": "~0.14.1",
+    "@angular-devkit/build-angular": "^20.3.7",
+    "@angular/animations": "^20.3.9",
+    "@angular/cli": "^20.3.7",
+    "@angular/common": "^20.3.9",
+    "@angular/compiler": "^20.3.9",
+    "@angular/core": "^20.3.9",
+    "@angular/forms": "^20.3.9",
+    "@angular/localize": "^20.3.9",
+    "@angular/platform-browser": "^20.3.9",
+    "@angular/platform-browser-dynamic": "^20.3.9",
+    "@angular/platform-server": "^20.3.9",
+    "@angular/router": "^20.3.9",
+    "@angular/ssr": "^20.3.7",
+    "@fortawesome/angular-fontawesome": "^3.0.0",
     "@fortawesome/fontawesome-common-types": "~6.7.2",
     "@fortawesome/fontawesome-svg-core": "~6.7.2",
     "@fortawesome/free-solid-svg-icons": "~6.7.2",
     "@mempool/mempool.js": "2.3.0",
-    "@ng-bootstrap/ng-bootstrap": "^16.0.0",
+    "@ng-bootstrap/ng-bootstrap": "^19.0.0",
     "@types/qrcode": "~1.5.0",
     "bootstrap": "~4.6.2",
     "browserify": "^17.0.0",
     "clipboard": "^2.0.11",
     "domino": "^2.1.6",
-    "echarts": "~5.6.0",
-    "ngx-echarts": "~17.2.0",
-    "ngx-infinite-scroll": "^17.0.0",
+    "echarts": "~5.4.0",
+    "ngx-echarts": "~20.0.2",
+    "ngx-infinite-scroll": "^20.0.0",
     "qrcode": "1.5.1",
     "rxjs": "~7.8.1",
     "esbuild": "^0.25.8",
     "tinyify": "^4.0.0",
     "tlite": "^0.1.9",
     "tslib": "~2.8.0",
-    "zone.js": "~0.14.4"
+    "zone.js": "~0.15.1"
   },
   "devDependencies": {
-    "@angular/compiler-cli": "^17.3.1",
-    "@angular/language-service": "^17.3.1",
-    "@types/node": "^18.11.9",
-    "@typescript-eslint/eslint-plugin": "^7.4.0",
-    "@typescript-eslint/parser": "^7.4.0",
-    "eslint": "^8.57.0",
+    "@angular/compiler-cli": "^20.3.9",
+    "@angular/language-service": "^20.3.9",
+    "@types/node": "^24.9.2",
+    "@typescript-eslint/eslint-plugin": "^8.46.2",
+    "@typescript-eslint/parser": "^8.46.2",
+    "eslint": "^9.39.0",
     "browser-sync": "^3.0.3",
     "http-proxy-middleware": "~2.0.6",
     "prettier": "^3.0.0",
     "source-map-support": "^0.5.21",
     "ts-node": "~10.9.1",
-    "typescript": "~5.4.3"
+    "typescript": "~5.8.3"
   },
   "optionalDependencies": {
     "@cypress/schematic": "^2.5.0",
