@@ -14,7 +14,8 @@ import { AddressInformation } from '@interfaces/node-api.interface';
 @Component({
   selector: 'app-address-group',
   templateUrl: './address-group.component.html',
-  styleUrls: ['./address-group.component.scss']
+  styleUrls: ['./address-group.component.scss'],
+  standalone: false,
 })
 export class AddressGroupComponent implements OnInit, OnDestroy {
   network = '';
