@@ -92,7 +92,8 @@ class AddressStats implements ChainStats {
 @Component({
   selector: 'app-address',
   templateUrl: './address.component.html',
-  styleUrls: ['./address.component.scss']
+  styleUrls: ['./address.component.scss'],
+  standalone: false,
 })
 export class AddressComponent implements OnInit, OnDestroy {
   network = '';
