@@ -8,7 +8,8 @@ import { AssetsService } from '@app/services/assets.service';
 @Component({
   selector: 'app-asset-group',
   templateUrl: './asset-group.component.html',
-  styleUrls: ['./asset-group.component.scss']
+  styleUrls: ['./asset-group.component.scss'],
+  standalone: false,
 })
 export class AssetGroupComponent implements OnInit {
   group$: Observable<any>;
