@@ -8,7 +8,8 @@ import { ServicesApiServices } from '@app/services/services-api.service';
 @Component({
   selector: 'app-bitcoin-invoice',
   templateUrl: './bitcoin-invoice.component.html',
-  styleUrls: ['./bitcoin-invoice.component.scss']
+  styleUrls: ['./bitcoin-invoice.component.scss'],
+  standalone: false,
 })
 export class BitcoinInvoiceComponent implements OnInit, OnChanges, OnDestroy {
   @Input() invoice;
