@@ -15,7 +15,8 @@ import { ServicesApiServices } from '@app/services/services-api.service';
 @Component({
   selector: 'app-block-preview',
   templateUrl: './block-preview.component.html',
-  styleUrls: ['./block-preview.component.scss']
+  styleUrls: ['./block-preview.component.scss'],
+  standalone: false,
 })
 export class BlockPreviewComponent implements OnInit, OnDestroy {
   network = '';
