@@ -8,7 +8,8 @@ import { StateService } from '@app/services/state.service';
   selector: 'app-fiat-selector',
   templateUrl: './fiat-selector.component.html',
   styleUrls: ['./fiat-selector.component.scss'],
-  changeDetection: ChangeDetectionStrategy.OnPush
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  standalone: false,
 })
 export class FiatSelectorComponent implements OnInit {
   fiatForm: UntypedFormGroup;
