@@ -21,7 +21,8 @@ function bestFitResolution(min, max, n): number {
 @Component({
   selector: 'app-mempool-block-view',
   templateUrl: './mempool-block-view.component.html',
-  styleUrls: ['./mempool-block-view.component.scss']
+  styleUrls: ['./mempool-block-view.component.scss'],
+  standalone: false,
 })
 export class MempoolBlockViewComponent implements OnInit, OnDestroy {
   autofit: boolean = false;
