@@ -10,7 +10,8 @@ import { AuthServiceMempool } from '@app/services/auth.service';
 @Component({
   selector: 'app-menu',
   templateUrl: './menu.component.html',
-  styleUrls: ['./menu.component.scss']
+  styleUrls: ['./menu.component.scss'],
+  standalone: false,
 })
 
 export class MenuComponent implements OnInit, OnDestroy {
