@@ -6,7 +6,8 @@ import { OpenGraphService } from '@app/services/opengraph.service';
 @Component({
   selector: 'app-privacy-policy',
   templateUrl: './privacy-policy.component.html',
-  styleUrls: ['./privacy-policy.component.scss']
+  styleUrls: ['./privacy-policy.component.scss'],
+  standalone: false,
 })
 export class PrivacyPolicyComponent {
   officialMempoolSpace = this.stateService.env.OFFICIAL_MEMPOOL_SPACE;
