@@ -3,7 +3,8 @@ import { WebsocketService } from '@app/services/websocket.service';
 
 @Component({
   selector: 'app-status-view',
-  templateUrl: './status-view.component.html'
+  templateUrl: './status-view.component.html',
+  standalone: false,
 })
 export class StatusViewComponent implements OnInit {
   constructor(
