@@ -31,6 +31,7 @@ interface MempoolStatsData {
   selector: 'app-dashboard',
   templateUrl: './dashboard.component.html',
   styleUrls: ['./dashboard.component.scss'],
+  standalone: false,
   changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class DashboardComponent implements OnInit, OnDestroy, AfterViewInit {
@@ -92,7 +93,7 @@ export class DashboardComponent implements OnInit, OnDestroy, AfterViewInit {
     private apiService: ApiService,
     private websocketService: WebsocketService,
     private seoService: SeoService,
-    @Inject(PLATFORM_ID) private platformId: Object,
+    @Inject(PLATFORM_ID) private platformId: object,
   ) {
     this.webGlEnabled = this.stateService.isBrowser && detectWebGL();
   }
