@@ -2,7 +2,8 @@ import { Directive, ElementRef, Renderer2 } from '@angular/core';
 import { environment } from '@environments/environment';
 
 @Directive({
-  selector: '[data-cy]'
+  selector: '[data-cy]',
+  standalone: false,
 })
 export class DataCyDirective {
   constructor(private el: ElementRef, private renderer: Renderer2) {
