@@ -9,7 +9,8 @@ import { FaqTemplateDirective } from '@app/docs/faq-template/faq-template.compon
 @Component({
   selector: 'app-api-docs',
   templateUrl: './api-docs.component.html',
-  styleUrls: ['./api-docs.component.scss']
+  styleUrls: ['./api-docs.component.scss'],
+  standalone: false,
 })
 export class ApiDocsComponent implements OnInit, AfterViewInit {
   private destroy$: Subject<any> = new Subject<any>();
