@@ -8,7 +8,8 @@ import { OpenGraphService } from '@app/services/opengraph.service';
 @Component({
   selector: 'app-docs',
   templateUrl: './docs.component.html',
-  styleUrls: ['./docs.component.scss']
+  styleUrls: ['./docs.component.scss'],
+  standalone: false,
 })
 export class DocsComponent implements OnInit {
 
