@@ -1,7 +1,8 @@
 import {  Directive, Input, TemplateRef } from '@angular/core';
 
 @Directive({
-  selector: 'ng-template[type]'
+  selector: 'ng-template[type]',
+  standalone: false,
 })
 export class FaqTemplateDirective {
   @Input() type: string;
