@@ -14,7 +14,8 @@ interface NodeGroup {
 @Component({
   selector: 'app-group-preview',
   templateUrl: './group-preview.component.html',
-  styleUrls: ['./group-preview.component.scss']
+  styleUrls: ['./group-preview.component.scss'],
+  standalone: false,
 })
 export class GroupPreviewComponent implements OnInit {
   nodes$: Observable<any>;
