@@ -4,7 +4,8 @@ import { AddressMatch, AddressTypeInfo } from '@app/shared/address-utils';
 @Component({
   selector: 'app-address-text',
   templateUrl: './address-text.component.html',
-  styleUrls: ['./address-text.component.scss']
+  styleUrls: ['./address-text.component.scss'],
+  standalone: false,
 })
 export class AddressTextComponent {
   @Input() address: string;
