@@ -4,7 +4,8 @@ import { AddressTypeInfo } from '@app/shared/address-utils';
 @Component({
   selector: 'app-address-type',
   templateUrl: './address-type.component.html',
-  styleUrls: []
+  styleUrls: [],
+  standalone: false,
 })
 export class AddressTypeComponent {
   @Input() address: AddressTypeInfo;
