@@ -11,7 +11,8 @@ export interface GeolocationData {
 @Component({
   selector: 'app-geolocation',
   templateUrl: './geolocation.component.html',
-  styleUrls: ['./geolocation.component.scss']
+  styleUrls: ['./geolocation.component.scss'],
+  standalone: false,
 })
 export class GeolocationComponent implements OnChanges {
   @Input() data: GeolocationData;
