@@ -5,7 +5,8 @@ import { StateService } from '@app/services/state.service';
 @Component({
   selector: 'app-sats',
   templateUrl: './sats.component.html',
-  styleUrls: ['./sats.component.scss']
+  styleUrls: ['./sats.component.scss'],
+  standalone: false,
 })
 export class SatsComponent implements OnInit {
   @Input() satoshis: number;
