@@ -2,13 +2,14 @@ import { Directive, TemplateRef, ViewContainerRef, Inject, PLATFORM_ID } from '@
 import { isPlatformServer } from '@angular/common';
 
 @Directive({
-  selector: '[serverOnly]'
+  selector: '[serverOnly]',
+  standalone: false,
 })
 export class ServerOnlyDirective {
   constructor(
     private templateRef: TemplateRef<any>,
     private viewContainer: ViewContainerRef,
-    @Inject(PLATFORM_ID) private platformId: Object
+    @Inject(PLATFORM_ID) private platformId: object
   ) {
     if (isPlatformServer(this.platformId)) {
       this.viewContainer.createEmbeddedView(this.templateRef);
