@@ -2,7 +2,8 @@ import { Pipe, PipeTransform } from '@angular/core';
 import { DomSanitizer, SafeHtml } from '@angular/platform-browser';
 
 @Pipe({
-  name: 'bitcoinsatoshis'
+  name: 'bitcoinsatoshis',
+  standalone: false,
 })
 export class BitcoinsatoshisPipe implements PipeTransform {
 
