@@ -1,6 +1,9 @@
 import { Pipe, PipeTransform } from '@angular/core';
 
-@Pipe({ name: 'ceil' })
+@Pipe({ 
+  name: 'ceil',
+  standalone: false,
+})
 export class CeilPipe implements PipeTransform {
   transform(nr: number) {
     return Math.ceil(nr);
