@@ -2,7 +2,8 @@ import { Pipe, PipeTransform } from '@angular/core';
 import { StateService } from '@app/services/state.service';
 
 @Pipe({
-  name: 'relativeUrl'
+  name: 'relativeUrl',
+  standalone: false,
 })
 export class RelativeUrlPipe implements PipeTransform {
 
